Fix integer conversion bug in `view_update_vy()`, for 32-bit platforms
The bug was introduced by commit 33d8677990f677f00d and resulted in the
cursor being placed as if the full `scroll-margin` value was always in
effect (i.e. placed on the line number corresponding to it), even for an
empty buffer.

This part of the code will be overhauled in the future, to avoid the
pitfalls of mixed type arithmetic, but for now this fix solves the
issue.

+++ src/view.c
@@ -95,7 +95,7 @@ static void view_update_vx(View *v)
 static void view_update_vy(View *v)
 {
     Window *window = v->window;
-    unsigned int margin = window_get_scroll_margin(window);
+    long margin = window_get_scroll_margin(window);
     long max_y = v->vy + window->edit_h - 1 - margin;
 
     if (v->cy < v->vy + margin) {
